Update README.md
To-Do List is a practical web application designed to help users manage their tasks efficiently. This project enables users to add and view tasks in a list format, check off tasks as completed, and delete tasks when no longer needed. The application utilizes the browser's local storage to persist task data, ensuring that the user's list remains intact even after closing or refreshing the browser. Built with HTML, CSS, and JavaScript, this To-Do List offers a user-friendly interface and reliable functionality for daily task management.